fix: :lipstick: Minor visual improvements
