+++ package.json
@@ -55,6 +55,9 @@
   "devDependencies": {
     "@types/react-animate-on-scroll": "^2.1.5",
     "@types/react-animations": "^1.0.0",
-    "gh-pages": "^3.2.3"
+    "env-cmd": "^10.1.0",
+    "gh-pages": "^3.2.3",
+    "prettier": "^2.5.1",
+    "prettier-plugin-organize-imports": "^2.3.4"
   }
 }
